feat: add static proxy in public plugin for local storage mode
