+++ src/index.ts
@@ -1,2 +1,3 @@
 export { default } from './service-api';
-export * from './service-api';
+export { default as publicPlugin } from './public-api';
+export * from './types';
